requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms==2.3.3
beautifulsoup4
black
isort
jupyter
matplotlib
networkx
notebook
numpy
opencv-contrib-python
pandas
pre-commit
pre-commit-hooks
scikit-learn
torch
torchaudio
torchvision
wordcloud
